Dynamic raw config options rendering
Since many of the parameters are rendered directly to the config as
strings, or lists of strings with a common key, it is easy to
consolidate all of the output logic into a single loop that renders
the options from a list of vars.

This has the added benefit for downstream consumers of allowing
them to inject additional config vars into the role and have them
rendered to the haproxy config without requiring a fork and
modification of the role itself.